+++ lib/jsdom/browser/Window.js
@@ -9,7 +9,6 @@ const Element = require("../living/generated/Element");
 const EventTarget = require("../living/generated/EventTarget");
 const PageTransitionEvent = require("../living/generated/PageTransitionEvent");
 const namedPropertiesWindow = require("../living/named-properties-window");
-const cssom = require("cssom");
 const postMessage = require("../living/post-message");
 const DOMException = require("domexception");
 const { btoa, atob } = require("abab");
@@ -28,15 +27,12 @@ const Storage = require("../living/generated/Storage");
 const createAbortController = require("../living/generated/AbortController").createInterface;
 const createAbortSignal = require("../living/generated/AbortSignal").createInterface;
 const reportException = require("../living/helpers/runtime-script-errors");
-const { matchesDontThrow } = require("../living/helpers/selectors");
 const { fireAnEvent } = require("../living/helpers/events");
 const SessionHistory = require("../living/window/SessionHistory");
+const { forEachSheetRuleOfElement, getResolvedValue, matches, properties } = require("../living/helpers/style-rules");
 
 const GlobalEventHandlersImpl = require("../living/nodes/GlobalEventHandlers-impl").implementation;
 const WindowEventHandlersImpl = require("../living/nodes/WindowEventHandlers-impl").implementation;
-
-const defaultStyleSheet = require("./default-stylesheet");
-let parsedDefaultStyleSheet;
 
 // NB: the require() must be after assigning `module.exports` because this require() is circular
 // TODO: this above note might not even be true anymore... figure out the cycle and document it, or clean up.
@@ -501,25 +497,8 @@ function Window(options) {
     elt = Element.convert(elt);
 
     const declaration = new CSSStyleDeclaration();
-    const { forEach, indexOf } = Array.prototype;
+    const { forEach } = Array.prototype;
     const { style } = elt;
-
-    function matches(rule, element) {
-      if (!rule.selectorText) {
-        return false;
-      }
-
-      const cssSelectorSplitRe = /((?:[^,"']|"[^"]*"|'[^']*')+)/;
-      const selectors = rule.selectorText.split(cssSelectorSplitRe);
-
-      for (const selectorText of selectors) {
-        if (selectorText !== "" && selectorText !== "," && matchesDontThrow(element, selectorText)) {
-          return true;
-        }
-      }
-
-      return false;
-    }
 
     function setPropertiesFromRule(rule) {
       if (matches(rule, elt)) {
@@ -533,113 +512,7 @@ function Window(options) {
       }
     }
 
-    if (!parsedDefaultStyleSheet) {
-      parsedDefaultStyleSheet = cssom.parse(defaultStyleSheet);
-    }
-
-    function forEachSheetRuleOfElement(element, handleRule) {
-      function handleSheet(sheet) {
-        forEach.call(sheet.cssRules, rule => {
-          if (rule.media) {
-            if (indexOf.call(rule.media, "screen") !== -1) {
-              forEach.call(rule.cssRules, handleRule);
-            }
-          } else {
-            handleRule(rule);
-          }
-        });
-      }
-
-      handleSheet(parsedDefaultStyleSheet);
-      forEach.call(element.ownerDocument.styleSheets, handleSheet);
-    }
-
     forEachSheetRuleOfElement(elt, setPropertiesFromRule);
-
-    /**
-     * properties for which `getResolvedValue` is implemented which is less than
-     * every supported property according to step 5
-     * @type Record<string, { inherited: boolean; initial: unknown; computedValue: 'as-specified' }>
-     */
-    const properties = {
-      visibility: {
-        inherited: true,
-        initial: "visible",
-        computedValue: "as-specified"
-      }
-    };
-
-    /**
-     * Naive implementation of https://www.w3.org/TR/CSS2/cascade.html#cascade
-     * based on the previous behavior of getComputedStyle
-     * Does not implement https://www.w3.org/TR/CSS2/cascade.html#specificity
-     * or rather specificity is only implemented by the order in which the matching
-     * rules appear. The last rule is the most specific while the first rule is
-     * the least specific.
-     * @param {Element} element
-     * @param {string} property
-     */
-    function getCascadedPropertyValue(element, property) {
-      let value = "";
-
-      forEachSheetRuleOfElement(elt, rule => {
-        if (matches(rule, element)) {
-          value = rule.style.getPropertyValue(property);
-        }
-      });
-
-      return value;
-    }
-
-    /**
-     * https://www.w3.org/TR/CSS2/cascade.html#specified-value
-     *
-     * @param {Element} element
-     * @param {string} property
-     */
-    function getSpecifiedValue(element, property) {
-      const cascade = getCascadedPropertyValue(element, property);
-
-      // 1.
-      if (cascade !== "") {
-        return cascade;
-      }
-
-      const { initial, inherited } = properties[property];
-      // 2
-      if (inherited === true && element.parentElement !== null) {
-        return getComputedValue(element.parentElement, property);
-      }
-
-      // 3
-      return initial;
-    }
-
-    /**
-     * https://drafts.csswg.org/css-cascade-4/#computed-value
-     * @param {Element} element
-     * @param {string} property
-     */
-    function getComputedValue(element, property) {
-      const { computedValue } = properties[property];
-      if (computedValue === "as-specified") {
-        return getSpecifiedValue(element, property);
-      }
-
-      throw new TypeError(`Unrecognized computed value instruction '${computedValue}'`);
-    }
-
-    /**
-     * https://drafts.csswg.org/cssom/#resolved-value
-     * Only implements `visibility`
-     * @param {Element} element
-     * @param {string} property
-     */
-    function getResolvedValue(element, property) {
-      // determined for special case properties none of which are implemented here
-      // so we skip to "any other property: The resolved value is the computed value."
-      return getComputedValue(element, property);
-    }
 
     // https://drafts.csswg.org/cssom/#dom-window-getcomputedstyle
     const declarations = Object.keys(properties);
